# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Garage SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/login:
    post:
      operationId: portalLogin
      tags: [portal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalLoginResult"
...
  /portal/change-password:
    post:
      operationId: portalChangePassword
      tags: [portal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
...
  /portal/orders/{id}:
    get:
      operationId: getPortalOrder
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /portal/orders/{id}/approve:
    post:
      operationId: portalApproveOrder
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Approval recorded
...
  /portal/vehicles/{id}/history:
    get:
      operationId: getPortalVehicleHistory
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle maintenance history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleHistory"
...
  /portal/reviews:
    post:
      operationId: createPortalReview
      tags: [portal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created

  # ── PUBLIC TRACKING ─────────────────────────────────────────────────────────
...
  /technicians:
    get:
      operationId: getTechnicians
      tags: [technicians]
      responses:
        "200":
          description: Technicians list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"

    post:
      operationId: createTechnician
      tags: [technicians]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"

    put:
      operationId: updateTechnician
      tags: [technicians]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}/orders:
    get:
      operationId: getTechnicianOrders
      tags: [technicians]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /technicians/{id}/stats:
    get:
      operationId: getTechnicianStats
      tags: [technicians]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianStats"

  # ── REVIEWS ───────────────────────────────────────────────────────────────────
...
  /superadmin/login:
    post:
      operationId: superAdminLogin
      tags: [superadmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuperAdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuperAdminLoginResult"
  /superadmin/tenants:
    get:
      operationId: getTenants
      tags: [superadmin]
      responses:
        "200":
          description: All tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"

    post:
      operationId: createTenant
      tags: [superadmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /superadmin/tenants/{id}:
    get:
      operationId: getTenant
      tags: [superadmin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

    put:
      operationId: updateTenant
      tags: [superadmin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
components:
  schemas:
...
    PortalLoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    PortalCustomer:
      type: object
      required: [id, fullName, phonePrimary]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phonePrimary:
          type: string
        isFirstLogin:
          type: boolean
        vehiclesCount:
          type: integer
    PortalLoginResult:
      type: object
      required: [customer, isFirstLogin]
      properties:
        customer:
          $ref: "#/components/schemas/PortalCustomer"
        isFirstLogin:
          type: boolean
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    ApprovalInput:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
        notes:
          type: string

    # ── CUSTOMER ───────────────────────────────────────────────────────────────
...
    Vehicle:
      type: object
      required: [id, customerId, plateNumber, make, model, year]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: ["string", "null"]
        bodyType:
          type: string
        fuelType:
          type: string
        transmission:
          type: string
        currentMileage:
          type: ["integer", "null"]
        engineSize:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    VehicleHistory:
      type: object
      required: [vehicle, orders, oilRecords]
      properties:
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        oilRecords:
          type: array
          items:
            $ref: "#/components/schemas/OilRecord"
...
    Order:
      type: object
      required: [id, orderNumber, vehicleId, customerId, status]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        vehicleId:
          type: integer
        customerId:
          type: integer
        status:
          type: string
        priority:
          type: string
        customerName:
          type: string
        vehiclePlate:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        estimatedCostMin:
          type: ["number", "null"]
        estimatedCostMax:
          type: ["number", "null"]
        grandTotal:
          type: ["number", "null"]
        receivedAt:
          type: string
        estimatedDeliveryAt:
          type: ["string", "null"]
        actualDeliveryAt:
          type: ["string", "null"]
        assignedTechnicianId:
          type: ["integer", "null"]
        assignedTechnicianName:
          type: ["string", "null"]
        paymentStatus:
          type: string
        customerComplaint:
          type: ["string", "null"]
        createdAt:
          type: string
...
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/OrderService"
            parts:
              type: array
              items:
                $ref: "#/components/schemas/OrderPart"
            photos:
              type: array
              items:
                $ref: "#/components/schemas/OrderPhoto"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusHistoryItem"
            invoice:
              $ref: "#/components/schemas/Invoice"
            mileageAtReception:
              type: ["integer", "null"]
            fuelLevel:
              type: ["string", "null"]
            diagnosisNotes:
              type: ["string", "null"]
            workDescription:
              type: ["string", "null"]
            internalNotes:
              type: ["string", "null"]
            customerApproved:
              type: ["boolean", "null"]
            trackingToken:
              type: ["string", "null"]
            totalPartsCost:
              type: ["number", "null"]
            totalLaborCost:
              type: ["number", "null"]
            discountAmount:
              type: ["number", "null"]
            taxRate:
              type: ["number", "null"]
...
    OrderService:
      type: object
      required: [id, orderId, serviceNameAr, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        serviceNameAr:
          type: string
        serviceNameEn:
          type: ["string", "null"]
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
        status:
          type: string
        warrantyDays:
          type: ["integer", "null"]
...
    OrderPart:
      type: object
      required: [id, orderId, partNameAr, quantity, unitCost, unitPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        partNameAr:
          type: string
        partNameEn:
          type: ["string", "null"]
        partNumber:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        quantity:
          type: number
        unitCost:
          type: number
        unitPrice:
          type: number
        warrantyDays:
          type: ["integer", "null"]
...
    OrderPhoto:
      type: object
      required: [id, orderId, photoUrl, type]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        photoUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        type:
          type: string
        captionAr:
          type: ["string", "null"]
        visibleToCustomer:
          type: boolean
        createdAt:
          type: string
    StatusHistoryItem:
      type: object
      required: [id, orderId, oldStatus, newStatus, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        oldStatus:
          type: string
        newStatus:
          type: string
        changedByName:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Invoice:
      type: object
      required: [id, invoiceNumber, orderId, total, status]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        vehicleInfo:
          type: string
        subtotal:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        totalPaid:
          type: number
        balanceDue:
          type: number
        currency:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        issuedAt:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        notesAr:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Technician:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        specialization:
          type: array
          items:
            type: string
        shift:
          type: string
        isAvailable:
          type: boolean
        ratingAverage:
          type: ["number", "null"]
        totalOrdersCompleted:
          type: integer
        activeOrdersCount:
          type: integer
    TechnicianInput:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        specialization:
          type: array
          items:
            type: string
        shift:
          type: string
    TechnicianUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        specialization:
          type: array
          items:
            type: string
        shift:
          type: string
        isAvailable:
          type: boolean
    TechnicianStats:
      type: object
      required: [totalCompleted, averageRating, totalRevenue]
      properties:
        totalCompleted:
          type: integer
        averageRating:
          type: number
        totalRevenue:
          type: number
        ordersThisMonth:
          type: integer

    # ── REVIEW ─────────────────────────────────────────────────────────────────
...
    ReviewInput:
      type: object
      required: [orderId, overallRating]
      properties:
        orderId:
          type: integer
        overallRating:
          type: integer
        speedRating:
          type: integer
        qualityRating:
          type: integer
        communicationRating:
          type: integer
        cleanlinessRating:
          type: integer
        commentAr:
          type: string
...
    SuperAdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        totpCode:
          type: string
    SuperAdminLoginResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        redirect:
          type: string
    Tenant:
      type: object
      required: [id, slug, nameAr, nameEn, plan, isActive]
      properties:
        id:
          type: integer
        slug:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        taglineAr:
          type: ["string", "null"]
        taglineEn:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        accentColor:
          type: string
        country:
          type: string
        currency:
          type: string
        plan:
          type: string
        maxUsers:
          type: integer
        maxOrdersPerMonth:
          type: integer
        subscriptionEnd:
          type: ["string", "null"]
        isActive:
          type: boolean
        phone:
          type: ["string", "null"]
        addressAr:
          type: ["string", "null"]
        createdAt:
          type: string
    TenantInput:
      type: object
      required: [slug, nameAr, nameEn, plan]
      properties:
        slug:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        taglineAr:
          type: string
        taglineEn:
          type: string
        country:
          type: string
        currency:
          type: string
        plan:
          type: string
        maxUsers:
          type: integer
        maxOrdersPerMonth:
          type: integer
        subscriptionEnd:
          type: string
        adminEmail:
          type: string
        adminPassword:
          type: string
        adminName:
          type: string
        phone:
          type: string
        addressAr:
          type: string
    TenantUpdate:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        taglineAr:
          type: string
        primaryColor:
          type: string
        accentColor:
          type: string
        plan:
          type: string
        maxUsers:
          type: integer
        maxOrdersPerMonth:
          type: integer
        subscriptionEnd:
          type: string
        isActive:
          type: boolean
        phone:
          type: string
        addressAr:
          type: string
...